requirements.txt
pymatgen
requests
mp-api
python-dotenv
